requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
python-dotenv==0.20.0
Werkzeug==2.2.3
SQLAlchemy==1.4.46
